fix(retex): trigger recomposition on text change
This commit fixes an issue where the Latex composable was not recomposing when the text property changed.

The `LaunchedEffect` key was changed from `textColor` to `text` to ensure that the composable is recomposed whenever the text to be rendered is updated.
Additionally, trailing commas were added for better code formatting.