feat: Update Room version to 2.7.0-rc03 and add Abiola field to SeriesEntity
This commit updates the Room library version to the release candidate 2.7.0-rc03.

It also introduces a new field `abiola` of type `Abiola` to the `SeriesEntity` and updates the database schema accordingly.
Additionally, a TypeConverter for the `Abiola` enum is added to handle database storage and retrieval and the `identityHash` in the schema file is updated.