feat: Update dependencies and project structure
This commit introduces dependency and project structure updates.

In `gradle/libs.versions.toml`, it adds version `4.1.1` for Koin and introduces the `koin-core` and `kotlinx-coroutines-core` dependencies.

In `app/build.gradle.kts`, it replaces a commented-out `library` project dependency with new module implementations for `database`, `jretex`, `model`, and `retex`, indicating a move to a more modular architecture.