ci: Remove CodeQL analysis workflow
This commit deletes the `codeql-analysis.yml` workflow file.

This workflow was configured to run on pushes and pull requests to the `develop` branch, as well as on a schedule. It performed CodeQL analysis for `java-kotlin` code by checking out the repository, initializing CodeQL, building the project, and then running the analysis.